// src/domain/enums/time-slot-period-enum.ts
export enum TimeSlotPeriod {
  PERIOD_NIGHT_1 = 'PERIOD_NIGHT_1', // 18:20 - 19:10
  PERIOD_NIGHT_2 = 'PERIOD_NIGHT_2', // 19:10 - 20:00
  PERIOD_NIGHT_3 = 'PERIOD_NIGHT_3', // 20:10 - 21:00
  PERIOD_NIGHT_4 = 'PERIOD_NIGHT_4', // 21:00 - 21:50
  PERIOD_NIGHT_5 = 'PERIOD_NIGHT_5', // 21:50 - 22:40
}