// src/domain/enums/subjetcs-enum.ts
export enum SubjectsEnum {
  ALGORITHMS = 'ALGORITHMS',
  DATA_STRUCTURES = 'DATA_STRUCTURES',
  DATABASES = 'DATABASES',
  SOFTWARE_ENGINEERING = 'SOFTWARE_ENGINEERING',
  OPERATING_SYSTEMS = 'OPERATING_SYSTEMS',
  COMPUTER_NETWORKS = 'COMPUTER_NETWORKS',
  INFORMATION_SYSTEMS = 'INFORMATION_SYSTEMS',
  ARTIFICIAL_INTELLIGENCE = 'ARTIFICIAL_INTELLIGENCE',
  WEB_DEVELOPMENT = 'WEB_DEVELOPMENT',
  MOBILE_DEVELOPMENT = 'MOBILE_DEVELOPMENT',
  PROJECT_MANAGEMENT = 'PROJECT_MANAGEMENT',
  HUMAN_COMPUTER_INTERACTION = 'HUMAN_COMPUTER_INTERACTION',
  CLOUD_COMPUTING = 'CLOUD_COMPUTING',
  CYBERSECURITY = 'CYBERSECURITY',
}